requirements.txt
streamlit
transformers
torch
docx
pdfminer.six
python-docx
requests
pandas